+++ tests/test_db.py
@@ -70,6 +70,13 @@ from environ.compat import DJANGO_POSTGRES
          '',
          '',
          ''),
+        ('sqlite:////full/path/to/your/file.sqlite',
+         'django.db.backends.sqlite3',
+         '/full/path/to/your/file.sqlite',
+         '',
+         '',
+         '',
+         ''),
         ('sqlite://:memory:',
          'django.db.backends.sqlite3',
          ':memory:',
@@ -94,6 +101,7 @@ from environ.compat import DJANGO_POSTGRES
         'cleardb',
         'mysql_no_password',
         'sqlite_empty',
+        'sqlite_file',
         'sqlite_memory',
         'ldap',
     ],
@@ -109,6 +117,53 @@ def test_db_parsing(url, engine, name, host, user, passwd, port):
         assert config['PASSWORD'] == passwd
         assert config['USER'] == user
         assert config['HOST'] == host
+
+
+def test_postgres_complex_db_name_parsing():
+    """Make sure we can use complex postgres host."""
+    env_url = (
+        'postgres://user:password@//cloudsql/'
+        'project-1234:us-central1:instance/dbname'
+    )
+
+    url = Env.db_url_config(env_url)
+
+    assert url['ENGINE'] == DJANGO_POSTGRES
+    assert url['HOST'] == '/cloudsql/project-1234:us-central1:instance'
+    assert url['NAME'] == 'dbname'
+    assert url['USER'] == 'user'
+    assert url['PASSWORD'] == 'password'
+    assert url['PORT'] == ''
+
+
+@pytest.mark.parametrize(
+    'scheme',
+    ['postgres', 'postgresql', 'psql', 'pgsql', 'postgis'],
+)
+def test_postgres_like_scheme_parsing(scheme):
+    """Verify all the postgres-like schemes parsed the same as postgres."""
+    env_url1 = (
+        'postgres://user:password@//cloudsql/'
+        'project-1234:us-central1:instance/dbname'
+    )
+    env_url2 = (
+        '{}://user:password@//cloudsql/'
+        'project-1234:us-central1:instance/dbname'
+    ).format(scheme)
+
+    url1 = Env.db_url_config(env_url1)
+    url2 = Env.db_url_config(env_url2)
+
+    assert url2['NAME'] == url1['NAME']
+    assert url2['PORT'] == url1['PORT']
+    assert url2['PASSWORD'] == url1['PASSWORD']
+    assert url2['USER'] == url1['USER']
+    assert url2['HOST'] == url1['HOST']
+
+    if scheme == 'postgis':
+        assert url2['ENGINE'] == 'django.contrib.gis.db.backends.postgis'
+    else:
+        assert url2['ENGINE'] == url1['ENGINE']
 
 
 def test_memory_sqlite_url_warns_about_netloc(recwarn):